requirements.txt
amqp==5.0.2
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dnspython==1.16.0
dominate==2.6.0
et-xmlfile==1.0.1
eventlet==0.30.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
greenlet==0.4.17
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
msgpack==1.0.1
numpy==1.19.5
openpyxl==3.0.5
pandas==1.1.5
prompt-toolkit==3.0.8
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
Rx==3.1.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.0
